id params edit

+++ public/js/postManager.js
@@ -1,17 +1,24 @@
 const delButton = async (e) => {
-  btnID = e.target.getAttribute("id");
+  const btnID = e.target.getAttribute("id");
+  typeBtn = e.target.innerText;
   if (btnID) {
-    const res = await fetch(`/api/post/${btnID}`, {
-      method: "DELETE",
-    });
+    if (typeBtn == "delete") {
+      const res = await fetch(`/api/post/${btnID}`, {
+        method: "DELETE",
+      });
 
-    if (res.ok) {
-      document.location.reload();
-    } else {
-      alert("Failed to delete post");
+      if (res.ok) {
+        document.location.reload();
+      } else {
+        alert("Failed to delete post");
+      }
+      res.ok ? alert("Post deleted") : alert("Failed to delete post");
+      return;
     }
-    res.ok ? alert("Post deleted") : alert("Failed to delete post");
+    document.location.replace(`/edit/${btnID}`);
   }
 };
+
+// const editButton
 
 document.querySelector(".mainpageposts").addEventListener("click", delButton);
